chore(day5): don't require reverse to parse binary number
Previously we reversed the string and summed all the indices (raised
to the power of two), where the bit was 1 to parse the numbers.

But we can of course use a more classical approach to it, traversing
our binary string and left shifting an accumulator by one for every
char.

Credits to @Hugsim for the inspiration!