chore(day5): don't split row and column
There is no need to split row and column up, as the only thing we do
are combining them to form a seat ID. This seat ID can be calculated
by considering the whole line as a binary number.

Once again credits goes to @Hugsim pointing this out to me.