feat: solve part 2 for day 11
Part 2 was really not that different from part 1, the main difference
was that we had to get the closest neigboring seats in each
direction, rather than the adjacent cells in the grid.

Since the closest neighboring seat in a direction can be found (if it
exists) by gradually moving in that direction it could be expressed in
a recursive manner.

Since Room.positions.get returns None if we call it with none-existing
positions it is harmless for us to make the coordinate types `i32`
instead of `usize`, even though `usize` is a tighter bound for the
values we'd like to express. However, by using `i32` we can safely
subtract from the coordinates, without having to check that we have an
overflow, since the cases which would yield a overflow subtraction are
the cases where we are at the borders of the grid, where the adjacent
nodes don't exist. Therefore we can ignore that "edge-case" and let
the HashMap deal with it for us.